Fix node 10 deprecation warnings
Prior to this, I was seeing:

(node:55150) [DEP0005] DeprecationWarning: Buffer() is deprecated due
to security and usability issues. Please use the Buffer.alloc(),
Buffer.allocUnsafe(), or Buffer.from() methods instead.

when running tests.

I left alone the calls to new Buffer in ModulesWithGlobalFallbackTest
and ModulesTest because there was a comment suggesting that they should
be testing deprecated apis.

See #3351